introduce point layer visualizer
introduce point layer visualizer for topLeft anchored and non-scaled features like points. polygons/lines are created in 0-1 space and scaled afterwards which doesn't work with points because it breaks prefab spawns
remove layer visualizer active property, it's easier to just remove it from vector module
add LayerType setting to vector layer object script. it isn't an option but a necessity because we cannot automatically decide which one is supposed to be used.